fix: 调整pageWrapper导出方式
