--- packages/core/src/platform/patch/navigation/common.js
@@ -1,15 +1,0 @@
-import { NavigationContainer, StackActions } from '@react-navigation/native'
-import PortalHost from '@mpxjs/webpack-plugin/lib/runtime/components/react/dist/mpx-portal/portal-host'
-import { useHeaderHeight } from '@react-navigation/elements'
-import { SafeAreaProvider, useSafeAreaInsets } from 'react-native-safe-area-context'
-import { GestureHandlerRootView } from 'react-native-gesture-handler'
-
-export {
-  NavigationContainer,
-  useHeaderHeight,
-  StackActions,
-  GestureHandlerRootView,
-  PortalHost,
-  SafeAreaProvider,
-  useSafeAreaInsets
-}
